backend/apps/bots/telethonecode/bots/telegram_bot/requirements.txt
python-telegram-bot==20.7
telethon==1.34.0
asyncpg==0.29.0
python-dotenv==1.0.0
g4f==0.4.9.9
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
